go-algorand v2.4.0-beta

+++ scripts/release/mule/sign/sign.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-# shellcheck disable=2035,2129,2162
+# shellcheck disable=2035,2129
 
 # TODO: This needs to be reworked a bit to support Darwin.
 
@@ -9,27 +9,55 @@ echo
 date "+build_release begin SIGN stage %Y%m%d_%H%M%S"
 echo
 
+if [ -z "$NETWORK" ]; then
+    echo "[$0] NETWORK is missing."
+    exit 1
+fi
+
+CHANNEL=$(./scripts/release/mule/common/get_channel.sh "$NETWORK")
 VERSION=${VERSION:-$(./scripts/compute_build_number.sh -f)}
-BRANCH=${BRANCH:-$(./scripts/compute_branch.sh)}
-CHANNEL=${CHANNEL:-$(./scripts/compute_branch_channel.sh "$BRANCH")}
 PKG_DIR="./tmp/node_pkgs"
 SIGNING_KEY_ADDR=dev@algorand.com
-
-# It seems that copying/mounting the gpg dir from another machine can result in insecure
-# access privileges, so set the correct permissions to avoid the following warning:
+OS_TYPE=$(./scripts/release/mule/common/ostype.sh)
+ARCHS=(amd64 arm arm64)
+ARCH_BITS=(x86_64 armv7l aarch64)
+# Note that we don't want to use $GNUPGHOME here because that is a documented env var for the gnupg
+# project and if it's set in the environment mule will automatically pick it up, which could have
+# unintended consequences and be hard to debug.
 #
-#   gpg: WARNING: unsafe permissions on homedir '/root/.gnupg'
-#
-find /root/.gnupg -type d -exec chmod 700 {} \;
-find /root/.gnupg -type f -exec chmod 600 {} \;
+# By naming it something other than $GNUPGHOME, it's essentially acting as an opt-in.
+GPG_DIR=${GPG_DIR:-/root/.gnupg}
 
-mkdir -p "$PKG_DIR"
-cd "$PKG_DIR"
+if ./scripts/release/mule/common/running_in_docker.sh
+then
+    # It seems that copying/mounting the gpg dir from another machine can result in insecure
+    # access privileges, so set the correct permissions to avoid the following warning:
+    #
+    #   gpg: WARNING: unsafe permissions on homedir '/root/.gnupg'
+    #
+    find "$GPG_DIR" -type d -exec chmod 700 {} \;
+    find "$GPG_DIR" -type f -exec chmod 600 {} \;
+fi
 
+# Note that when downloading from the cloud that we'll get all packages for all architectures.
 if [ -n "$S3_SOURCE" ]
 then
-    aws s3 cp --recursive --exclude "*" --include "*$CHANNEL*$VERSION*" "s3://$S3_SOURCE/$CHANNEL/$VERSION" .
+    i=0
+    for arch in "${ARCHS[@]}"; do
+        arch_bit="${ARCH_BITS[$i]}"
+        (
+            mkdir -p "$PKG_DIR/$OS_TYPE/$arch"
+            cd "$PKG_DIR"
+            # Note the underscore after ${arch}!
+            # Recall that rpm packages have the arch bit in the filenames (i.e., "x86_64" rather than "amd64").
+            # Also, the order of the includes/excludes is important!
+            aws s3 cp --recursive --exclude "*" --include "*${arch}_*" --include "*$arch_bit.rpm" --exclude "*.sig" --exclude "*.asc" --exclude "*.asc.gz" "s3://$S3_SOURCE/$CHANNEL/$VERSION" .
+        )
+        i=$((i + 1))
+    done
 fi
+
+cd "$PKG_DIR"
 
 # TODO: "$PKG_TYPE" == "source"
 
@@ -41,37 +69,42 @@ OS_TYPES=($(find . -mindepth 1 -maxdepth 1 -type d -printf '%f\n'))
 for os in "${OS_TYPES[@]}"; do
     if [ "$os" = linux ]
     then
-        ARCHS=(amd64 arm arm64)
         for arch in "${ARCHS[@]}"; do
-            (
-                mkdir -p "$os/$arch"
-                cd "$os/$arch"
+            if [ -d "$os/$arch" ]
+            then
+                # Only do the subsequent operations in a subshell if the directory is not empty.
+                if stat -t "$os/$arch/"* > /dev/null 2>&1
+                then
+                (
+                    cd "$os/$arch"
 
-                # Clean package directory of any previous operations.
-                rm -rf hashes* *.sig *.asc *.asc.gz
+                    # Clean package directory of any previous operations.
+                    rm -rf hashes* *.sig *.asc *.asc.gz
 
-                for file in *.tar.gz *.deb
-                do
-                    gpg -u "$SIGNING_KEY_ADDR" --detach-sign "$file"
-                done
+                    for file in *.tar.gz *.deb
+                    do
+                        gpg -u "$SIGNING_KEY_ADDR" --detach-sign "$file"
+                    done
 
-                for file in *.rpm
-                do
-                    gpg -u rpm@algorand.com --detach-sign "$file"
-                done
+                    for file in *.rpm
+                    do
+                        gpg -u rpm@algorand.com --detach-sign "$file"
+                    done
 
-                HASHFILE="hashes_${CHANNEL}_${os}_${arch}_${VERSION}"
-                md5sum *.tar.gz *.deb *.rpm >> "$HASHFILE"
-                shasum -a 256 *.tar.gz *.deb *.rpm >> "$HASHFILE"
-                shasum -a 512 *.tar.gz *.deb *.rpm >> "$HASHFILE"
+                    HASHFILE="hashes_${CHANNEL}_${os}_${arch}_${VERSION}"
+                    md5sum *.tar.gz *.deb *.rpm >> "$HASHFILE"
+                    shasum -a 256 *.tar.gz *.deb *.rpm >> "$HASHFILE"
+                    shasum -a 512 *.tar.gz *.deb *.rpm >> "$HASHFILE"
 
-                gpg -u "$SIGNING_KEY_ADDR" --detach-sign "$HASHFILE"
-                gpg -u "$SIGNING_KEY_ADDR" --clearsign "$HASHFILE"
+                    gpg -u "$SIGNING_KEY_ADDR" --detach-sign "$HASHFILE"
+                    gpg -u "$SIGNING_KEY_ADDR" --clearsign "$HASHFILE"
 
-                STATUSFILE="build_status_${CHANNEL}_${os}-${arch}_${VERSION}"
-                gpg -u "$SIGNING_KEY_ADDR" --clearsign "$STATUSFILE"
-                gzip -c "$STATUSFILE.asc" > "$STATUSFILE.asc.gz"
-            )
+                    STATUSFILE="build_status_${CHANNEL}_${os}-${arch}_${VERSION}"
+                    gpg -u "$SIGNING_KEY_ADDR" --clearsign "$STATUSFILE"
+                    gzip -c "$STATUSFILE.asc" > "$STATUSFILE.asc.gz"
+                )
+                fi
+            fi
         done
     fi
 done
